+++ webviews/components/MakeGradient.svelte
@@ -1,51 +1,41 @@
 <script lang="ts">
     import { randomColorGenerator } from "../utils/randomColorGenerator";
     import { copyCss } from "../utils/copyCss";
-    
-    // Select options 
+
+    let colorStops: ColorStop[] = [
+        { color: randomColorGenerator(), position: 0 },
+        { color: randomColorGenerator(), position: 100 },
+    ];
+
+    // Select options
     const linearDirections = [
-        {angle: 0, logo: "⬆️"},
-        {angle: 45, logo: "↗️"},
-        {angle: 270, logo: "➡️"},
-        {angle: 135, logo: "↘️"},
-        {angle: 180, logo: "⬇️"},
-        {angle: 225, logo: "↙️"},
-        {angle: 90, logo: "⬅️"},
-        {angle: 315, logo: "↖️"},
+        { angle: 0, logo: "⬆️" },
+        { angle: 45, logo: "↗️" },
+        { angle: 270, logo: "➡️" },
+        { angle: 135, logo: "↘️" },
+        { angle: 180, logo: "⬇️" },
+        { angle: 225, logo: "↙️" },
+        { angle: 90, logo: "⬅️" },
+        { angle: 315, logo: "↖️" },
     ];
 
     const radialPositions = [
-        {position: "center", logo: "🅾️"},
-        {position: "top", logo: "⬆️"},
-        {position: "right top", logo: "↗️"},
-        {position: "right", logo: "➡️"},
-        {position: "right bottom", logo: "↘️"},
-        {position: "bottom", logo: "⬇️"},
-        {position: "left bottom", logo: "↙️"},
-        {position: "left", logo: "⬅️"},
-        {position: "left top", logo: "↖️"}
+        { position: "center", logo: "🅾️" },
+        { position: "top", logo: "⬆️" },
+        { position: "right top", logo: "↗️" },
+        { position: "right", logo: "➡️" },
+        { position: "right bottom", logo: "↘️" },
+        { position: "bottom", logo: "⬇️" },
+        { position: "left bottom", logo: "↙️" },
+        { position: "left", logo: "⬅️" },
+        { position: "left top", logo: "↖️" },
     ];
-    
+
     // CSS code to copy
     let code: HTMLTextAreaElement;
 
-    // Defines the gradient colors. Initially random
-    let colors = {
-        primaryColor: randomColorGenerator(), 
-        secondaryColor: randomColorGenerator()
-    };
-
-    // Custom types for selected options
-    type SelectedDirection = {
-        angle?: number;
-    };
-
-    type SelectedPosition = {
-        position?: string;
-    };
-
-    let selectedAngle: SelectedDirection = {angle: 90};
-    let selectedPosition: SelectedPosition = {position: "center"}; 
+    let selectedAngle: SelectedDirection = { angle: 90 };
+    let selectedPosition: SelectedPosition = { position: "center" };
 
     let animate = false;
 
@@ -53,17 +43,76 @@
 
     let animationLength = 1;
 
+    function addColor() {
+        // Temporary limit on the number of color stops, may change in the future
+        if (colorStops.length > 5) return;
+
+        const newColor = {
+            color: randomColorGenerator(),
+            position:
+                // Make sure position value doesn't go over 100
+                colorStops[colorStops.length - 1].position + 10 > 100
+                    ? colorStops[colorStops.length - 1].position / 2
+                    : colorStops[colorStops.length - 1].position + 10,
+        };
+
+        colorStops = [...colorStops, newColor];
+    }
+
+    function removeColor(index: number) {
+        if (colorStops.length === 2) {
+            return;
+        }
+
+        colorStops = colorStops.filter((_, i) => i !== index);
+    }
+
+    let gradient: string;
+
+    $: {
+        // Creating a copy to prevent the UI position changing
+        // after the ColorStop object is sorted
+        const colorStopsCopy = [...colorStops];
+        const colors = colorStopsCopy
+            // Construct gradient code here
+            .sort((a, b) => a.position - b.position)
+            .map((color, i) => `${color.color} ${color.position}%`)
+            .join(", ");
+
+        const mode = radial ? "radial-gradient(circle at " : "linear-gradient(";
+
+        const direction = radial
+            ? selectedPosition.position
+            : `${selectedAngle.angle}deg`;
+
+        const animation = animate
+            ? `\nbackground-size: 400% 400%;\nanimation: animationName ${animationLength}s ease infinite;
+                \n@keyframes animationName {
+                    \n  0% { background-position: 0% 50%; }
+                    \n  50% { background-position: 100% 50%; }
+                    \n  100% { background-position: 0% 50%; }
+                  \n}`
+            : "";
+        gradient = `background: ${mode}${direction}, ${colors}); ${animation}`;
+    }
+
+    function randomise() {
+        for (let i = 0; i < colorStops.length; i++) {
+            colorStops[i].color = randomColorGenerator();
+        }
+    }
+
 </script>
 
 <div class="container">
     <div class="selections">
-        Gradient Type
+        <span class="select-heading">Gradient Type</span>
         <select bind:value={radial}>
             <option value={false} selected>Linear</option>
-            <option value={true}>Radial</option>        
+            <option value={true}>Radial</option>
         </select>
-        
-        Gradient Direction
+
+        <span class="select-heading">Gradient Direction</span>
         {#if radial}
             <select bind:value={selectedPosition}>
                 {#each radialPositions as position}
@@ -81,153 +130,67 @@
                 {/each}
             </select>
         {/if}
-        
-        Animate Gradient
+
+        <span class="select-heading">Animate Gradient</span>
         <select bind:value={animate}>
             <option value={false} selected>❌</option>
-            <option value={true}>✅</option>        
+            <option value={true}>✅</option>
         </select>
 
-        Animation Length
-        <input disabled={!animate} type="range" min=1 max=59 bind:value={animationLength} />
+        <span class="select-heading">Animation Length</span>
+        <input
+            disabled={!animate}
+            type="range"
+            min="1"
+            max="59"
+            bind:value={animationLength}
+        />
+    </div>
+    <div class="gradient-preview" style={gradient} />
+
+    <div class="top-btn">
+        <button on:click={addColor}>Add Color</button>
+
+        <button on:click={randomise}>Randomize</button>
     </div>
 
-    <div class="primaryColor">
-        <input type="color" bind:value={colors.primaryColor} />        
-        <input type="text" bind:value={colors.primaryColor}/>
+    <div class="color-stops">
+        {#each colorStops as color, i}
+            <div class="color-control">
+                <input type="color" bind:value={color.color} />
+                <input
+                    class="color-input"
+                    type="number"
+                    min="0"
+                    max="100"
+                    bind:value={color.position}
+                />
+                <button class="delete-btn" on:click={() => removeColor(i)}>
+                    <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        fill="#fff"
+                        viewBox="0 0 16 16"
+                        width="16"
+                        height="16"
+                        style="text-align:center"
+                        ><path
+                            d="M3.72 3.72a.75.75 0 0 1 1.06 0L8 6.94l3.22-3.22a.749.749 0 0 1 1.275.326.749.749 0 0 1-.215.734L9.06 8l3.22 3.22a.749.749 0 0 1-.326 1.275.749.749 0 0 1-.734-.215L8 9.06l-3.22 3.22a.751.751 0 0 1-1.042-.018.751.751 0 0 1-.018-1.042L6.94 8 3.72 4.78a.75.75 0 0 1 0-1.06Z"
+                        /></svg
+                    >
+                </button>
+            </div>
+        {/each}
     </div>
-
-    <div class="secondaryColor">
-        <input type="color" bind:value={colors.secondaryColor} />
-        <input type="text" bind:value={colors.secondaryColor}/>
-    </div>
-
-    <div class="randomBtn">
-        <button on:click={() => {
-            colors.primaryColor = randomColorGenerator();
-            colors.secondaryColor = randomColorGenerator();
-        }}>
-            Randomize Colors
-        </button>
-    </div>
-    <!-- Animated linear gradient -->
-    {#if animate && !radial}
-        <div class="animateDisplay" 
-            style="background-image: linear-gradient({selectedAngle.angle}deg, 
-            {colors.primaryColor}, {colors.secondaryColor}); --animationLength:{animationLength}s"
-        />
-    <!-- Animated radial gradient -->
-    {:else if animate && radial}
-        <div class="animateDisplay" 
-            style="background-image: radial-gradient(circle at {selectedPosition.position}, 
-            {colors.primaryColor}, {colors.secondaryColor}); --animationLength:{animationLength}s"
-        />
-    <!-- Static radial gradient -->
-    {:else if !animate && radial}
-        <div class="noAnimateDisplay" 
-            style="background-image: radial-gradient(circle at {selectedPosition.position}, 
-            {colors.primaryColor}, {colors.secondaryColor})"
-        />
-    <!-- Static linear gradient -->
-    {:else}
-        <div class="noAnimateDisplay" 
-            style="background-image: linear-gradient({selectedAngle.angle}deg, 
-            {colors.primaryColor}, {colors.secondaryColor})"
-        />
-    {/if}
 
     <div class="code">
-        <button on:click={() => { copyCss(code) }}>
+        <button
+            class="copy-btn"
+            on:click={() => {
+                copyCss(code);
+            }}
+        >
             Copy CSS
         </button>
-        <!-- This is the gradient code to display to users -->
-        {#if animate && !radial}                
-            <textarea readonly bind:this={code}>
-                background: linear-gradient({selectedAngle.angle}deg, {colors.primaryColor}, {colors.secondaryColor});
-                animation: gradient {animationLength}s ease infinite;
-                @keyframes gradient &lbrace;
-                    0%&lbrace;background-position:0% 50%&rbrace;
-                    50%&lbrace;background-position:100% 50%&rbrace;
-                    100%&lbrace;background-position:0% 50%&rbrace;
-                &rbrace;
-            </textarea>
-        {:else if animate && radial}
-            <textarea readonly bind:this={code}>
-                background: radial-gradient(circle at {selectedPosition.position}, {colors.primaryColor}, {colors.secondaryColor});
-                animation: gradient {animationLength}s ease infinite;
-                @keyframes gradient &lbrace;
-                    0%&lbrace;background-position:0% 50%&rbrace;
-                    50%&lbrace;background-position:100% 50%&rbrace;
-                    100%&lbrace;background-position:0% 50%&rbrace;
-                &rbrace;
-            </textarea>
-        {:else if !animate && radial}
-            <textarea readonly bind:this={code}>
-                background: radial-gradient(circle at {selectedPosition.position}, {colors.primaryColor}, {colors.secondaryColor});
-            </textarea>
-        {:else}
-            <textarea readonly bind:this={code}>
-                background: linear-gradient({selectedAngle.angle}deg, {colors.primaryColor}, {colors.secondaryColor});
-            </textarea>
-        {/if}
+        <textarea readonly bind:this={code}>{gradient}</textarea>
     </div>
 </div>
-
-<style>
-    /* Using css grid layout*/
-    .container {
-        display: grid;
-        grid-template-columns: repeat(5, 1fr);
-        grid-template-rows: 0.5fr repeat(5, 1fr);
-        grid-column-gap: 0px;
-        grid-row-gap: 0px;
-    }
-
-    .selections {
-        grid-area: 1 / 1 / 2 / 6;
-        margin: auto;
-    }
-
-    .primaryColor {
-        grid-area: 5 / 1 / 6 / 3; 
-    }
-
-    .secondaryColor {
-        grid-area: 5 / 4 / 6 / 6; 
-        margin-left: auto;    
-    }
-
-    .randomBtn {
-        grid-area: 5 / 3 / 6 / 4; 
-        margin-top: 0.5rem;
-    }
-
-    .code {
-        grid-area: 6 / 1 / 7 / 6;
-    }
-
-    .noAnimateDisplay {
-        grid-area: 2 / 1 / 5 / 6;
-        border: 2px solid rgba(1,1,1,0,0.35);
-        height: 350px;
-        margin-bottom: 0.5rem;
-        margin-top: 0.5rem;
-    }
-
-    .animateDisplay {
-        grid-area: 2 / 1 / 5 / 6;
-        border: 2px solid rgba(1,1,1,0,0.35);
-        height: 350px;
-        background-size: 400% 400%;
-        animation: gradient var(--animationLength) ease infinite;
-        margin-bottom: 0.5rem;
-        margin-top: 0.5rem;
-    }
-
-    @keyframes gradient {
-        0%{background-position:0% 50%}
-        50%{background-position:100% 50%}
-        100%{background-position:0% 50%}
-    }
-            
-</style>